bench/identity-perf: rename from bench/perf + reference results on main
Two-part change on the identity-agent perf harness:

* Rename bench/perf/ -> bench/identity-perf/ for symmetry with the new
  bench/context-perf/ harness (added in the follow-up commit). Only
  mechanical rewrites: go.mod module path, cmd/*/main.go router
  import, Dockerfile COPY paths, docker-compose.yml Dockerfile refs,
  README + run.sh + run-scaling.sh + .gitignore + router.go comments.
  No behavior change.

* Append a Reference results section to README with the full-matrix
  sweep from a clean origin/main checkout on an AMD EPYC 9254 (24
  cores / 48 threads, 384 GiB DDR5, Ubuntu 24.04 + Docker 29.1.3).
  Cross-checked against the PR #413 baseline on the same hardware:
  fcap-only regressed 15-25% and fcap-audience 30-40% at CPU-bound
  configs between #413 baseline and current main. non_2xx / errors
  are 0 in both datasets so this isn't a correctness change reducing
  work — per-request cost went up. Follow-up: bisect the post-#413
  identityagent commits (3f8d1a6 audience fail-closed, 103cbdb
  union-read shadow-mode fallback are the natural suspects).
