+++ bench/identity-perf/.gitignore
@@ -1,5 +1,5 @@
 results/
-# Anchor to the bench/perf root — otherwise these also match the
+# Anchor to the bench/identity-perf root — otherwise these also match the
 # cmd/{configserver,seeder,loadgen} source directories.
 /configserver
 /seeder
