+++ bench/identity-perf/cmd/seeder/main.go
@@ -21,7 +21,7 @@ import (
 	"sync"
 	"time"
 
-	"github.com/adcontextprotocol/adcp-go/bench/perf/internal/router"
+	"github.com/adcontextprotocol/adcp-go/bench/identity-perf/internal/router"
 	"github.com/adcontextprotocol/adcp-go/targeting/identityhash"
 	"github.com/redis/go-redis/v9"
 )
